backend/requirements.txt
python-telegram-bot[webhooks]
python-telegram-bot==20.3
httpx
aiohttp
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
pydantic[email]
python-jose
passlib[bcrypt]
python-dotenv
greenlet
requests
